% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaverted.R
\name{qaverted}
\alias{qaverted}
\title{Get Number Averted}
\usage{
qaverted(E, E_n, u, I)
}
\arguments{
\item{E}{the original vacine effectiveness}

\item{E_n}{the new vacine effectiveness}

\item{u}{the uptake rate in the population}

\item{I}{the current incidence in the population}
}
\value{
A the number of averted events
}
\description{
Calculate the number of averted events after increasing vaccine effectiveness in the population.
}
\examples{
# In a population changing vaccine effectiveness from
# 0.6 to 0.7 where the uptake rate is 0.5 and there is an incidence of
# 1 million yearly
qaverted(0.6, 0.7, 0.5, 1e6)
}
